StateV2: Remove Observer lifetime constraints
Prior to this commit, StateV2 internally tracked which states were
observed during a given `memo`/`effect` function by storing any such
observed states in a `MutableSet` (called `observed`), and then, after
the function returns, the actual long-term `dependencies` and
`dependents` lists would be updated based on these observations.
This was done so we do not need to shift around a ton of lists before
each re-evaluation to clear all previous dependencies, only to re-add
most of them when the function re-observes them.

This commit gets rid of the intermediate `MutableSet`. The
dependents/dependencies lists are now updated immediately on
observation.
As a result, it is now possible to observe states even after the main
`memo`/`effect` function has already returned.
This makes cases like `memo { Modifier.effect { MyEffect(myState()) }
}`, which previously would not function correctly, now fully supported.

It achives this without compromising on the reason for why we had the
intermediate `MutableSet` to begin with, it instead just employs a
different optimization strategy which should generally have equal, if
not better, performance. See the docs on the new `Edge` type for
details.

Additionally, this commit introduces a bunch of unit tests for the
StateV2 implementation, because it was previously severely lacking these
and it's not immediately obvious if some detail is broken, as general
functionality might be unaffected in most other cases.

Source-Commit: d0a120eb52d5846af9c81fb96337594b6a9c0510